Initial Commit
This is the first commit for the internal Chain Blast 2 repository.
It is a mess, and the game has some features implemented already, but mostly half-baked with no UI. SkilletUI will come at some point to help with that!

This is a version of OpenSMCE from between Beta 4.8.0 and Beta 4.9.0 releases. I don't remember the exact commit I've copied it from, but this is an amalgamation of a lot of that, the actual Chain Blast 2 code (most of which has been reused from the original Chain Blast) and some assets.

Congratulations on reading this information, I know that this is the worst possible code organization where you just copy code from another repository so you have to maintain two copies! I am just a single person, so expect that a lot and happy going through all the trash and garbage where nothing makes sense! :D

But in all seriousness, if you're here from the future, and you're still reading this somehow, thank you so much! I don't even know what this will become in the future! Maybe nothing, or maybe everything! Greetings from the past! :D

This commit adds some third-party libraries and assets used in OpenSMCE. They use their respective licenses. More info on that will be added when there will actually be someone who would read that information. I don't want to bother with typing it out right now!